+++ otterdog/providers/github/__init__.py
@@ -57,7 +57,7 @@ class Github:
                        repo_name: str,
                        path: str,
                        content: str,
-                       message: Optional[str] = None) -> None:
+                       message: Optional[str] = None) -> bool:
         return self.rest_client.update_content(org_id, repo_name, path, content, message)
 
     def get_org_settings(self, org_id: str, included_keys: set[str], no_web_ui: bool) -> dict[str, Any]:
@@ -160,3 +160,6 @@ class Github:
 
     def get_ref_for_pull_request(self, org_id: str, repo_name: str, pull_number: str) -> str:
         return self.rest_client.get_ref_for_pull_request(org_id, repo_name, pull_number)
+
+    def sync_from_template_repository(self, org_id: str, repo_name: str, template_repository: str) -> list[str]:
+        return self.rest_client.sync_from_template_repository(org_id, repo_name, template_repository)